api/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
prometheus-client==0.19.0
httpx==0.26.0
